fix: decoration errors of experimentalDecorators
decoration errors of experimentalDecorators
