+++ src/index.ts
@@ -101,7 +101,6 @@ export * from "./components/anim/OAnimationEvent"
 export * from "./components/anim/curveAnim/AnimationMonitor"
 export * from "./components/anim/curveAnim/AttributeAnimCurve"
 export * from "./components/anim/curveAnim/PropertyAnimClip"
-export * from "./components/anim/curveAnim/PropertyAnimDecoration"
 export * from "./components/anim/curveAnim/PropertyAnimation"
 export * from "./components/anim/curveAnim/PropertyAnimationEvent"
 export * from "./components/anim/curveAnim/PropertyHelp"
